requirements.txt
Flask==3.0.0
gunicorn==21.2.0
flask-cors==4.0.0
flask-caching==2.1.0
flask-limiter==3.5.0
google-cloud-bigquery==3.13.0
